Update Project_3.qmd
This is great information for your data dictionary. Why not make a separate CSV file? It is something that you will sometimes want to refer to to see quickly and easily what the variables are. 

Also for .qmd, you have to add two spaces after each line if you want it to appear as a separate line (more table like). Otherwise the next line will start immedately in one big paragraph. I added spaces so you should be able to render it again and see the info on separate lines